deleted file AddUser.js, route /add-user, created folder Student_Page and moved related filed in it
