+++ frontend/src/Components/Student_Page/Body.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useContext, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import Add from "../Add_User/Components/Add";
-import UserProfileButtons from "../Add_User/Components/UserProfileButtons";
-import Card from "../Components/Card";
-import { fetchStudent } from "../services/utils";
-import { ShowSkills, AddSkill } from "./ShowAddSkill";
-import { AdminContext } from "../App";
+import Add from "../../Add_User/Components/Add";
+import UserProfileButtons from "../../Add_User/Components/UserProfileButtons";
+import Card from "../Card";
+import { fetchStudent } from "../../services/utils";
+import { ShowSkills, AddSkill } from "../ShowAddSkill";
+import { AdminContext } from "../../App";
 // import "./Body.css";
 
 function Body({ studentDetails }) {
