+++ frontend/src/Components/Student_Page/Footer.jsx
